Refactor GameView into mixins
