+++ tienlen_gui/__init__.py
@@ -56,7 +56,8 @@ from .overlays import (
     GameOverOverlay,
 )
 from .view import GameView, main
-from .hud import HUDPanel
+from .hud import HUDPanel, HUDMixin
+from .overlay_manager import OverlayMixin
 
 __all__ = [
     'TABLE_THEMES', 'PLAYER_COLORS', 'HAND_SPACING', 'HORIZONTAL_MARGIN',
@@ -72,6 +73,6 @@ __all__ = [
     'InGameMenuOverlay', 'SettingsOverlay', 'GameSettingsOverlay', 'GraphicsOverlay',
     'AudioOverlay', 'RulesOverlay', 'HowToPlayOverlay', 'TutorialOverlay', 'SavePromptOverlay',
     'ProfileOverlay', 'GameOverOverlay', 'GameView', 'main',
-    'HUDPanel',
+    'HUDPanel', 'HUDMixin', 'OverlayMixin',
     'SHADOW_OFFSET', 'SHADOW_BLUR', 'SHADOW_ALPHA'
 ]
